refactor: move response models into separate module
