+++ cargo-shuttle/src/client.rs
@@ -11,9 +11,9 @@ use reqwest_retry::policies::ExponentialBackoff;
 use reqwest_retry::RetryTransientMiddleware;
 use serde::de::DeserializeOwned;
 use serde::Deserialize;
-use shuttle_common::models::{deployment, project, user};
+use shuttle_common::models::{deployment, log, project, secret, service, user};
 use shuttle_common::project::ProjectName;
-use shuttle_common::{log, secret, service, ApiKey, ApiUrl};
+use shuttle_common::{ApiKey, ApiUrl};
 use tokio::net::TcpStream;
 use tokio_tungstenite::tungstenite::client::IntoClientRequest;
 use tokio_tungstenite::{connect_async, MaybeTlsStream, WebSocketStream};
