+++ common/Cargo.toml
@@ -17,26 +17,13 @@ rocket = "0.5.0-rc.2" # todo: remove this dep and debug the error from shuttle_s
 rustrict = "0.5.0"
 semver = { version = "1.0.13", optional = true }
 serde = "1.0.137"
-serde_json = "1.0.83"
+serde_json = { version = "1.0.85", optional = true }
 strum = { version = "0.24.1", features = ["derive"] }
 tracing = "0.1.36"
 uuid = { version = "1.1.1", features = ["v4", "serde"] }
 
 [features]
-default = []
-full = [
-    "log",
-    "secret",
-    "service",
-    "version",
-    "models"
-]
+default = ["models", "version"]
 
-models = ["display", "chrono"]
-
-display = ["crossterm", "comfy-table"]
+models = ["comfy-table", "chrono", "crossterm", "serde_json"]
 version = ["anyhow", "semver", "regex"]
-
-log = ["crossterm"]
-secret = ["display"]
-service = ["display"]
