+++ common/src/models/log.rs
