+++ common/src/models/mod.rs
@@ -1,3 +1,7 @@
 pub mod deployment;
+pub mod log;
 pub mod project;
+pub mod resource;
+pub mod secret;
+pub mod service;
 pub mod user;
