+++ deployer/Cargo.toml
@@ -42,7 +42,6 @@ uuid = { version = "1.1.2", features = ["v4"] }
 
 [dependencies.shuttle-common]
 version = "0.5.2"
-features = ["full"]
 path = "../common"
 
 [dependencies.shuttle-proto]
