+++ deployer/src/handlers/mod.rs
@@ -9,7 +9,8 @@ use axum::{extract::BodyStream, Json};
 use chrono::{TimeZone, Utc};
 use fqdn::FQDN;
 use futures::TryStreamExt;
-use shuttle_common::{log, secret, LogItem};
+use shuttle_common::models::{log, secret};
+use shuttle_common::LogItem;
 use tower_http::auth::RequireAuthorizationLayer;
 use tower_http::trace::TraceLayer;
 use tracing::{debug, debug_span, error, field, Span};
@@ -76,7 +77,7 @@ pub fn make_router(
 
 async fn list_services(
     Extension(persistence): Extension<Persistence>,
-) -> Result<Json<Vec<shuttle_common::service::Response>>> {
+) -> Result<Json<Vec<shuttle_common::models::service::Response>>> {
     let services = persistence
         .get_all_services()
         .await?
@@ -90,7 +91,7 @@ async fn list_services(
 async fn get_service(
     Extension(persistence): Extension<Persistence>,
     Path((_project_name, service_name)): Path<(String, String)>,
-) -> Result<Json<shuttle_common::service::Detailed>> {
+) -> Result<Json<shuttle_common::models::service::Detailed>> {
     if let Some(service) = persistence.get_service_by_name(&service_name).await? {
         let deployments = persistence
             .get_deployments(&service.id)
@@ -111,7 +112,7 @@ async fn get_service(
             .map(Into::into)
             .collect();
 
-        let response = shuttle_common::service::Detailed {
+        let response = shuttle_common::models::service::Detailed {
             name: service.name,
             deployments,
             resources,
@@ -128,7 +129,7 @@ async fn get_service_summary(
     Extension(persistence): Extension<Persistence>,
     Extension(proxy_fqdn): Extension<FQDN>,
     Path((project_name, service_name)): Path<(String, String)>,
-) -> Result<Json<shuttle_common::service::Summary>> {
+) -> Result<Json<shuttle_common::models::service::Summary>> {
     if let Some(service) = persistence.get_service_by_name(&service_name).await? {
         let deployment = persistence
             .get_active_deployment(&service.id)
@@ -141,7 +142,7 @@ async fn get_service_summary(
             .map(Into::into)
             .collect();
 
-        let response = shuttle_common::service::Summary {
+        let response = shuttle_common::models::service::Summary {
             uri: format!("https://{}.{proxy_fqdn}", project_name),
             name: service.name,
             deployment,
@@ -191,7 +192,7 @@ async fn delete_service(
     Extension(persistence): Extension<Persistence>,
     Extension(deployment_manager): Extension<DeploymentManager>,
     Path((_project_name, service_name)): Path<(String, String)>,
-) -> Result<Json<shuttle_common::service::Detailed>> {
+) -> Result<Json<shuttle_common::models::service::Detailed>> {
     if let Some(service) = persistence.get_service_by_name(&service_name).await? {
         let old_deployments = persistence
             .delete_deployments_by_service_id(&service.id)
@@ -216,7 +217,7 @@ async fn delete_service(
 
         persistence.delete_service(&service.id).await?;
 
-        let response = shuttle_common::service::Detailed {
+        let response = shuttle_common::models::service::Detailed {
             name: service.name,
             deployments: old_deployments.into_iter().map(Into::into).collect(),
             resources,
@@ -321,7 +322,7 @@ async fn build_logs_websocket_handler(mut s: WebSocket, persistence: Persistence
 
     for log in backlog.into_iter().filter_map(Log::into_build_log_stream) {
         match (log.state, log.message) {
-            (shuttle_common::deployment::State::Building, Some(msg)) => {
+            (shuttle_common::models::deployment::State::Building, Some(msg)) => {
                 let sent = s.send(ws::Message::Text(msg)).await;
                 last_timestamp = log.timestamp;
 
@@ -330,9 +331,9 @@ async fn build_logs_websocket_handler(mut s: WebSocket, persistence: Persistence
                     return;
                 }
             }
-            (shuttle_common::deployment::State::Building, None) => {}
-            (shuttle_common::deployment::State::Queued, _)
-            | (shuttle_common::deployment::State::Built, _) => {}
+            (shuttle_common::models::deployment::State::Building, None) => {}
+            (shuttle_common::models::deployment::State::Queued, _)
+            | (shuttle_common::models::deployment::State::Built, _) => {}
             _ => {
                 debug!("closing channel after reaching more than just build logs");
                 let _ = s.close().await;
@@ -345,7 +346,7 @@ async fn build_logs_websocket_handler(mut s: WebSocket, persistence: Persistence
         if log.id == id && log.timestamp > last_timestamp {
             if let Some(log) = persistence::Log::from(log).into_build_log_stream() {
                 match (log.state, log.message) {
-                    (shuttle_common::deployment::State::Building, Some(msg)) => {
+                    (shuttle_common::models::deployment::State::Building, Some(msg)) => {
                         let sent = s.send(ws::Message::Text(msg)).await;
 
                         // Client disconnected?
@@ -353,9 +354,9 @@ async fn build_logs_websocket_handler(mut s: WebSocket, persistence: Persistence
                             return;
                         }
                     }
-                    (shuttle_common::deployment::State::Queued, _)
-                    | (shuttle_common::deployment::State::Built, _) => {}
-                    (shuttle_common::deployment::State::Building, None) => {}
+                    (shuttle_common::models::deployment::State::Queued, _)
+                    | (shuttle_common::models::deployment::State::Built, _) => {}
+                    (shuttle_common::models::deployment::State::Building, None) => {}
                     _ => break,
                 }
             }
