+++ proto/Cargo.toml
@@ -11,7 +11,6 @@ tonic = "0.8.0"
 
 [dependencies.shuttle-common]
 version = "0.5.2"
-features = ["full"]
 path = "../common"
 
 [build-dependencies]
