+++ service/Cargo.toml
@@ -55,6 +55,7 @@ uuid = { version = "1.1.2", features = ["v4"] }
 
 [dependencies.shuttle-common]
 version = "0.5.2"
+default-features = false
 path = "../common"
 
 [features]
